requirements.txt
aiogram~=2.20
python-dotenv~=0.20.0
psycopg2-binary ~= 2.9.3